+++ internal/api/handlers/task/api.go
@@ -1,0 +1,14 @@
+package task
+
+import (
+	"example.com/m/v2/internal/api/handlers"
+
+	"github.com/gin-gonic/gin"
+)
+
+func NewHandler(router *gin.Engine, service handlers.MainUseCase) {
+	apiTask := router.Group("/tasks")
+	{
+		apiTask.POST("/", AddTask(service))
+	}
+}