+++ internal/api/handlers/user/user_test.go
@@ -1,0 +1,10 @@
+package user
+
+import (
+	"testing"
+)
+
+func TestCreateUser(t *testing.T) {
+	t.Parallel()
+	
+}