+++ internal/api/handlers/CHANGELOG.md
@@ -7,6 +7,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Create unit test case
-- Generate token
-- Make setup mock
+- Get token from cookie
+- Make handle `UseCase`